+++ src/ProgramLoader.cpp
@@ -90,7 +90,7 @@ private:
         if (sectionExpression_.indexIn(QString::fromUtf8(line)) == -1)
             return false;
 
-        uint16_t startAddress = static_cast<uint16_t>(sectionExpression_.cap(2).toInt(nullptr, 16));
+        int32_t startAddress = sectionExpression_.cap(2).toInt(nullptr, 16);
         if (startAddress < programStartAddress_)
         {
             programStartAddress_ = startAddress;
@@ -131,7 +131,7 @@ private:
         if (addrExpression_.indexIn(QString::fromUtf8(tokens[0])) != -1)
         {
             sl.address = addrExpression_.cap(2).toInt(nullptr, 16);
-            uint16_t relAddr = static_cast<uint16_t>(sl.address - programStartAddress_);
+            int32_t relAddr = sl.address - programStartAddress_;
             QByteArray data = QByteArray::fromHex(addrExpression_.cap(3).toUtf8());
 
             if (binaryData_.length() < relAddr + data.length())
@@ -163,7 +163,7 @@ private:
     int run_{0};
     bool needMoreRuns_{false};
     QByteArray binaryData_;
-    uint16_t programStartAddress_{std::numeric_limits<uint16_t>::max()};
+    int32_t programStartAddress_{std::numeric_limits<uint16_t>::max()};
     bool hasProgramStartAddress_{false};
     QList<Program::SourceLine> sourceLines_;
 };
